+++ core/src/main/kotlin/org/evomaster/core/problem/api/ApiWsAction.kt
@@ -1,6 +1,6 @@
 package org.evomaster.core.problem.api
 
-import org.evomaster.core.problem.api.auth.AuthenticationInfo
+import org.evomaster.core.problem.enterprise.auth.AuthenticationInfo
 import org.evomaster.core.problem.api.param.Param
 import org.evomaster.core.search.action.Action
